+++ src/main/java/de/jeff_media/angelchest/config/Permissions.java
@@ -22,4 +22,5 @@ public final class Permissions {
     public static final String ADMIN = "angelchest.admin";
     public static final String LIST = "angelchest.list";
     public static final String GUI = "angelchest.gui";
+    public static final String PREVENT_VICTIM_CHEST = "angelchest.preventvictimchest";
 }